+++ template-parts/loops/search-row.php
@@ -7,8 +7,8 @@
  */
 
 ?>
-<tr>
-	<td style="border-bottom: 2px solid #FF5F00">
+<tr class="webbooks-search-row">
+	<td>
 		<h5>
 			<a href="<?php the_permalink(); ?>" class="card-title">
 				<?php the_title(); ?>